Correct coAuthors interpretation; Changes to Author.toString()
Apparently, the publication’s HTML page changes whether there is an
active login session or not.
If there is, the correct cssQuery for the coauthors is
"a.publication-author-name”. If not, it is
"meta[name=citation_author]”. This is the reason why it was being
returned empty - it is working correctly now.